Rename math to algebra.linear
