+++ apsg/algebra/linear/scalar.py
