+++ apsg/algebra/linear/tensor.py
