+++ apsg/core.py
@@ -15,7 +15,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from apsg.setting import settings
-from apsg.math.helper import (
+from apsg.algebra.linear.helper import (
     KentDistribution,
     sind,
     cosd,
