--- apsg/math/__init__.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-"""
-Matrix, Vector and Tensor algebra in low dimension.
-"""
-
-
-__all__ = (
-	"Tensor",
-	"Matrix2", "Matrix3", "Matrix4", "MatrixError",
-	"Vector2", "Vector3", "Vector4", "VectorError",
-	)
-
-
-from apsg.math.tensor import Tensor
-from apsg.math.matrix import Matrix2, Matrix3, Matrix4, MatrixError
-from apsg.math.vector import Vector2, Vector3, Vector4, VectorError
-from apsg.math.helper import sind, cosd, tand, asind, acosd # TODO Re-import other functions?
